requirements.txt
torch>=2.0.0
torchvision
numpy
pyyaml
tqdm
pandas
jiwer
tensorboard